Reorganizes sources
